add pages and handle routers
