+++ src/pages/About.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function About() {
+  return (
+    <div>
+      <h1>About</h1>
+      <p>This is the about page</p>
+    </div>
+  );
+}
+export default About;
